eternityready/requirements.txt
-e git+https://github.com/django-nonrel/django@nonrel-1.5#egg=django 
argparse==1.2.1
django-mongodb-engine==0.6.0
djangotoolbox==1.8.0
pymongo==3.3.0
sqlparse==0.2.1
wsgiref==0.1.2
